+++ lib/models/ClientStatic.d.ts
@@ -3,5 +3,5 @@ import { Config } from './Config';
 export interface ClientStatic {
     defaultClientConfig: Config;
     create(config?: Config): ContensisClient;
-    configure(config: Config): any;
+    configure(config: Config): void;
 }