+++ src/models/IUserOperations.ts
@@ -1,4 +1,4 @@
-import { User, UserCreate } from './User';
+import { User, UserToCreate } from './User';
 import { PagedList } from 'contensis-core-api';
 import { UserListOptions } from './UserListOptions';
 import { Group } from './Group';
@@ -12,7 +12,7 @@ export interface IUserOperations {
     getByEmail(email: string): Promise<User>;
     list(options?: UserListOptions): Promise<PagedList<User>>;
     getUserGroups(userIdOrOptions: string | UserGroupsOptions): Promise<PagedList<Group>>;
-    create(user: UserCreate, suspended?: boolean): Promise<User>;
+    create(user: UserToCreate, suspended?: boolean): Promise<User>;
     update(user: User): Promise<User>;
     updatePassword(options: UserUpdatePasswordOptions): Promise<void>;
     delete(userId: string): Promise<void>;