+++ src/models/User.ts
@@ -35,4 +35,4 @@ export interface User {
     passwordChanged: Date;
 }
 
-export type UserCreate = Omit<User, 'id'> & Partial<Pick<User, 'id'>>;
+export type UserToCreate = Omit<User, 'id'> & Partial<Pick<User, 'id'>>;