+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2022-08-16
+### Changes
+- users: added *UserToCreate* type for creating a user to support optional id
+- groups: added *GroupToCreate* type for creating a group to support optional id
+- upgraded to TypeScript 4
+
 ## [2.0.9] - 2022-07-08
 ### Added
 - users: added pagination to *getUserGroups* operation